# Make `indent` faster by preallocating the output string