On my system, it took just under 2 microseconds to indent a ~2 KB text with 50 lines:

    % cargo criterion --bench indent
    indent                  time:   [1.8829 us 1.8891 us 1.8948 us]

Preallocating the output string makes the function about 20% faster:

    % cargo criterion --bench indent
    indent                  time:   [1.4647 us 1.4670 us 1.4698 us]
                            change: [-23.426% -22.816% -22.328%] (p = 0.00 < 0.05)
                            Performance has improved.